Adding authentication and authorization (#1)

+++ FoodRecipeAppAPI/Controllers/IngredientsController.cs
@@ -1,9 +1,12 @@
-﻿using FoodRecipeAppAPI.Data.Dtos.Ingredients;
+﻿using FoodRecipeAppAPI.Auth.Models;
+using FoodRecipeAppAPI.Data.Dtos.Ingredients;
 using FoodRecipeAppAPI.Data.Entities;
 using FoodRecipeAppAPI.Data.Repositories;
 using FoodRecipeAppAPI.Data.Repositories.Ingredients;
 using FoodRecipeAppAPI.Data.Repositories.Recipes;
+using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
+using System.IdentityModel.Tokens.Jwt;
 
 namespace FoodRecipeAppAPI.Controllers
 {
@@ -14,12 +17,14 @@ namespace FoodRecipeAppAPI.Controllers
         private readonly IIngredientsRepository _ingredientsRepository;
         private readonly IRecipesRepository _recipesRepository;
         private readonly ICategoriesRepository _categoriesRepository;
+        private readonly IAuthorizationService _authorizationService;
 
-        public IngredientsController(IIngredientsRepository ingredientsRepository, IRecipesRepository recipesRepository, ICategoriesRepository categoriesRepository)
+        public IngredientsController(IIngredientsRepository ingredientsRepository, IRecipesRepository recipesRepository, ICategoriesRepository categoriesRepository, IAuthorizationService authorizationService)
         {
             _ingredientsRepository = ingredientsRepository;
             _recipesRepository = recipesRepository;
             _categoriesRepository = categoriesRepository;
+            _authorizationService = authorizationService;
         }
 
         [HttpGet]
@@ -87,6 +92,7 @@ namespace FoodRecipeAppAPI.Controllers
         }
 
         [HttpPost]
+        [Authorize(Roles = AppRoles.User)]
         public async Task<ActionResult<IngredientDto>> CreateAsync(int categoryId, int recipeId, CreateIngredientDto dto)
         {
             var category = await _categoriesRepository.GetAsync(categoryId);
@@ -96,11 +102,25 @@ namespace FoodRecipeAppAPI.Controllers
                 return NotFound();
             }
 
+            var categoryAuthorizationResult = await _authorizationService.AuthorizeAsync(User, category, PolicyNames.ResourceOwner);
+
+            if (!categoryAuthorizationResult.Succeeded)
+            {
+                return Forbid();
+            }
+
             var recipe = await _recipesRepository.GetAsync(recipeId);
 
-            if (recipe == null)
+            if (recipe == null) 
             {
                 return NotFound();
+            }
+
+            var recipeAuthorizationResult = await _authorizationService.AuthorizeAsync(User, recipe, PolicyNames.ResourceOwner);
+
+            if (!recipeAuthorizationResult.Succeeded)
+            {
+                return Forbid();
             }
 
             var ingredient = new Ingredient
@@ -109,7 +129,8 @@ namespace FoodRecipeAppAPI.Controllers
                 Type = dto.Type,
                 IsVegetarian = dto.IsVegetarian,
                 IsVegan = dto.IsVegan,
-                Recipe = recipe
+                Recipe = recipe,
+                UserId = User.FindFirst(JwtRegisteredClaimNames.Sub).Value
             };
 
             await _ingredientsRepository.CreateAsync(ingredient);
@@ -126,6 +147,7 @@ namespace FoodRecipeAppAPI.Controllers
 
         [HttpPut]
         [Route("{ingredientId}")]
+        [Authorize(Roles = AppRoles.User)]
         public async Task<ActionResult<IngredientDto>> UpdateAsync(int categoryId, int recipeId, int ingredientId, CreateIngredientDto dto)
         {
             var category = await _categoriesRepository.GetAsync(categoryId);
@@ -149,6 +171,13 @@ namespace FoodRecipeAppAPI.Controllers
                 return NotFound();
             }
 
+            var authorizationResult = await _authorizationService.AuthorizeAsync(User, ingredient, PolicyNames.ResourceOwner);
+
+            if (!authorizationResult.Succeeded)
+            {
+                return Forbid();
+            }
+
             ingredient.Name = dto.Name;
             ingredient.Type = dto.Type;
             ingredient.IsVegetarian = dto.IsVegetarian;
@@ -168,6 +197,7 @@ namespace FoodRecipeAppAPI.Controllers
 
         [HttpDelete]
         [Route("{ingredientId}")]
+        [Authorize(Roles = AppRoles.User)]
         public async Task<ActionResult> DeleteAsync(int categoryId, int recipeId, int ingredientId)
         {
             var category = await _categoriesRepository.GetAsync(categoryId);
@@ -189,6 +219,13 @@ namespace FoodRecipeAppAPI.Controllers
             if (ingredient == null)
             {
                 return NotFound();
+            }
+
+            var authorizationResult = await _authorizationService.AuthorizeAsync(User, ingredient, PolicyNames.ResourceOwner);
+
+            if (!authorizationResult.Succeeded)
+            {
+                return Forbid();
             }
 
             await _ingredientsRepository.DeleteAsync(ingredient);
